requirements.txt
fastapi
uvicorn
pypdf
python-docx
sentence-transformers
numpy
faiss-cpu
openai
pinecone-client
requests
python-multipart
